+++ tests/test_cover.py
@@ -139,7 +139,9 @@ async def test_ceipblind_dr(
 ) -> None:
     """Test CeIpBlind DIN Rail."""
     central, mock_client, _ = central_client_factory
-    cover: CeIpBlind = cast(CeIpBlind, helper.get_prepared_custom_entity(central, "VCU7807849", 2))
+    cover: CeIpBlind = cast(
+        CeIpBlind, helper.get_prepared_custom_entity(central, "VCU7807849", 14)
+    )
     assert cover.usage == EntityUsage.CE_PRIMARY
     assert cover.service_method_names == (
         "close",
@@ -157,7 +159,7 @@ async def test_ceipblind_dr(
     assert cover.is_closed is True
     await cover.set_position(position=81)
     assert mock_client.method_calls[-1] == call.set_value(
-        channel_address="VCU7807849:2",
+        channel_address="VCU7807849:14",
         paramset_key=ParamsetKey.VALUES,
         parameter="COMBINED_PARAMETER",
         value="L2=0,L=81",
@@ -166,48 +168,48 @@ async def test_ceipblind_dr(
     # test unconfirmed values
     assert cover._e_level.unconfirmed_last_value_send == 0.81
     assert cover._e_level_2.unconfirmed_last_value_send == _CLOSED_LEVEL
-    await central.event(const.INTERFACE_ID, "VCU7807849:2", "LEVEL", 0.81)
-    await central.event(const.INTERFACE_ID, "VCU7807849:2", "LEVEL_2", _CLOSED_LEVEL)
+    await central.event(const.INTERFACE_ID, "VCU7807849:14", "LEVEL", 0.81)
+    await central.event(const.INTERFACE_ID, "VCU7807849:14", "LEVEL_2", _CLOSED_LEVEL)
     assert cover._e_level.unconfirmed_last_value_send is None
     assert cover._e_level_2.unconfirmed_last_value_send is None
 
-    await central.event(const.INTERFACE_ID, "VCU7807849:1", "LEVEL", 0.81)
-    await central.event(const.INTERFACE_ID, "VCU7807849:1", "LEVEL_2", _CLOSED_LEVEL)
+    await central.event(const.INTERFACE_ID, "VCU7807849:13", "LEVEL", 0.81)
+    await central.event(const.INTERFACE_ID, "VCU7807849:13", "LEVEL_2", _CLOSED_LEVEL)
     assert cover.current_position == 81
     assert cover.is_closed is False
     await cover.open()
     assert mock_client.method_calls[-1] == call.set_value(
-        channel_address="VCU7807849:2",
+        channel_address="VCU7807849:14",
         paramset_key=ParamsetKey.VALUES,
         parameter="COMBINED_PARAMETER",
         value="L2=100,L=100",
     )
     assert cover._e_level.unconfirmed_last_value_send == _OPEN_LEVEL
     assert cover._e_level_2.unconfirmed_last_value_send == _OPEN_TILT_LEVEL
-    await central.event(const.INTERFACE_ID, "VCU7807849:1", "LEVEL", _OPEN_LEVEL)
-    await central.event(const.INTERFACE_ID, "VCU7807849:1", "LEVEL_2", _OPEN_TILT_LEVEL)
-    await central.event(const.INTERFACE_ID, "VCU7807849:2", "LEVEL", _OPEN_LEVEL)
-    await central.event(const.INTERFACE_ID, "VCU7807849:2", "LEVEL_2", _OPEN_TILT_LEVEL)
+    await central.event(const.INTERFACE_ID, "VCU7807849:13", "LEVEL", _OPEN_LEVEL)
+    await central.event(const.INTERFACE_ID, "VCU7807849:13", "LEVEL_2", _OPEN_TILT_LEVEL)
+    await central.event(const.INTERFACE_ID, "VCU7807849:14", "LEVEL", _OPEN_LEVEL)
+    await central.event(const.INTERFACE_ID, "VCU7807849:14", "LEVEL_2", _OPEN_TILT_LEVEL)
     assert cover._e_level.unconfirmed_last_value_send is None
     assert cover._e_level_2.unconfirmed_last_value_send is None
     assert cover.current_position == 100
     assert cover.current_tilt_position == 100
     await cover.close()
     assert mock_client.method_calls[-1] == call.set_value(
-        channel_address="VCU7807849:2",
+        channel_address="VCU7807849:14",
         paramset_key=ParamsetKey.VALUES,
         parameter="COMBINED_PARAMETER",
         value="L2=0,L=0",
     )
-    await central.event(const.INTERFACE_ID, "VCU7807849:1", "LEVEL", _CLOSED_LEVEL)
+    await central.event(const.INTERFACE_ID, "VCU7807849:13", "LEVEL", _CLOSED_LEVEL)
     assert cover.is_opening is None
     assert cover.is_closing is None
-    await central.event(const.INTERFACE_ID, "VCU7807849:1", "ACTIVITY_STATE", 1)
+    await central.event(const.INTERFACE_ID, "VCU7807849:13", "ACTIVITY_STATE", 1)
     assert cover.is_opening is True
-    await central.event(const.INTERFACE_ID, "VCU7807849:1", "ACTIVITY_STATE", 2)
+    await central.event(const.INTERFACE_ID, "VCU7807849:13", "ACTIVITY_STATE", 2)
     assert cover.is_closing is True
 
-    await central.event(const.INTERFACE_ID, "VCU7807849:1", "LEVEL", 0.5)
+    await central.event(const.INTERFACE_ID, "VCU7807849:13", "LEVEL", 0.5)
     assert cover._channel_level == 0.5
     assert cover.current_position == 50