+++ changelog.md
@@ -3,7 +3,8 @@
 - Add central link methods to click event
 - Add reportValueUsage, addLink, removeLink and getLinks to client
 - Add version to code
-- Move operation_mode to channel
+- Fix wrong channel assignment for HmIP-DRBLI4
+- Add operation_mode to channel
 
 # Version 2024.10.8 (2024-10-15)